requirements.txt
# Core
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# Medical imaging
pydicom>=2.4.0

# OpenAI API
openai>=1.0.0

# Visualization
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
